chore: move retry tests into separate test namespace
