+++ tests/Integration/Momento.Sdk.Tests/Retry/utils/TestRetryMetricsCollector.cs
@@ -1,7 +1,8 @@
 using Momento.Sdk.Config.Retry;
 using System.Collections.Generic;
 
-namespace Momento.Sdk.Tests.Integration.Cache;
+// namespace Momento.Sdk.Tests.Integration.Cache;
+namespace Momento.Sdk.Tests.Integration.Retry;
 
 public class TestRetryMetricsCollector
 {
